+++ src/Modules/Profile/components/BananaList.vue
@@ -1,13 +1,13 @@
 <template>
   <div id="bananas">
     <div v-if="!bananaCount" class="my-1">
-      {{ $i18n('profile.banana.none', { name: recipientName }) }}
+      {{ $i18n('profile.banana.none', { name: recipient.name }) }}
     </div>
 
     <div v-if="canGiveBanana && !hasGivenBanana" class="mb-2">
       <div v-if="showTextarea">
         <b-alert variant="success" show>
-          {{ $i18n('profile.banana.details', { name: recipientName }) }}
+          {{ $i18n('profile.banana.details', { name: recipient.name }) }}
           <br>
           <strong>
             {{ $i18n('profile.banana.undo') }}
@@ -41,7 +41,7 @@
             :disabled="!canSendBanana"
             @click="trySendBanana"
           >
-            {{ $i18n('profile.banana.give', { name: recipientName }) }}
+            {{ $i18n('profile.banana.give', { name: recipient.name }) }}
           </b-button>
         </div>
       </div>
@@ -51,7 +51,7 @@
           size="sm"
           @click="toggleTextarea"
         >
-          {{ $i18n('profile.banana.give', { name: recipientName }) }}
+          {{ $i18n('profile.banana.give', { name: recipient.name }) }}
         </b-button>
       </div>
     </div>
@@ -59,11 +59,11 @@
     <BananaListEntry
       v-for="b in bananaList"
       :key="b.id"
-      :author="{...b, avatar: b.photo }"
-      :created-at="b.createdAt"
-      :text="b.msg"
+      :author="b.user"
+      :created-at="b.time"
+      :text="b.message"
       :can-remove="canRemoveBanana"
-      :recipient-id="recipientId"
+      :recipient-id="recipient.id"
       @close-dialog="closeDialog"
     />
   </div>
@@ -72,10 +72,9 @@
 <script>
 import $ from 'jquery'
 
-import { sendBanana } from '@/api/profile'
+import { sendBanana } from '@/api/banana'
 import i18n from '@/helper/i18n'
 import { pulseError, pulseInfo } from '@/script'
-import DataUser from '@/stores/user'
 
 import BananaListEntry from './BananaListEntry'
 import { HTTP_RESPONSE } from '@/consts'
@@ -83,8 +82,7 @@ import { HTTP_RESPONSE } from '@/consts'
 export default {
   components: { BananaListEntry },
   props: {
-    recipientId: { type: Number, required: true },
-    recipientName: { type: String, required: true },
+    recipient: { type: Object, required: true },
     canGiveBanana: { type: Boolean, default: false },
     canRemoveBanana: { type: Boolean, default: false },
     bananas: { type: Array, default: () => { return [] } },
@@ -109,12 +107,7 @@ export default {
   methods: {
     async trySendBanana () {
       try {
-        await sendBanana(this.recipientId, this.bananaText.trim())
-
-        // Fake reactive update by inserting submitted data into the UI
-        const fakeBanana = this.getFakeBanana()
-        this.bananaList.unshift(fakeBanana)
-        this.bananaCount += 1
+        this.bananaList.unshift(await sendBanana(this.recipient.id, this.bananaText.trim()))
 
         // Reset UI and component state
         pulseInfo(i18n('profile.banana.sent'))
@@ -126,7 +119,7 @@ export default {
         if (err.code === HTTP_RESPONSE.BAD_REQUEST) {
           pulseError(i18n('profile.banana.messageTooShort'))
         } else if (err.code === HTTP_RESPONSE.FORBIDDEN) {
-          pulseError(i18n('profile.banana.alreadyGiven', { name: this.recipientName }))
+          pulseError(i18n('profile.banana.alreadyGiven', { name: this.recipient.name }))
         } else {
           console.error(err)
           pulseError(i18n('error_unexpected'))
@@ -136,14 +129,6 @@ export default {
     toggleTextarea () {
       this.showTextarea = !this.showTextarea
       $.fancybox.update()
-    },
-    getFakeBanana () {
-      return {
-        createdAt: new Date().toISOString(),
-        id: DataUser.getters.getUserId(),
-        photo: DataUser.getters.getAvatar(),
-        msg: this.bananaText.trim(),
-      }
     },
     closeDialog () {
       $.fancybox.close()