+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 ## Refactoring
 - Added an enum for wall types in the backend #1996 !3573 @alex.simm
+- Bananas now use the API !3581 @AntonBallmaier
 
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm